reformat code compile flags: set warning level, warning treat as errors

+++ main.cpp
@@ -23,21 +23,17 @@
 #*                                                              *
 #\* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
+
 #include <math.h>
 #include <stdlib.h>
 
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__)
-// MsWindows-on ez is kell
 #include <windows.h>
 #endif // Win32 platform
 
 #include <GL/gl.h>
 #include <GL/glu.h>
-// A GLUT-ot le kell tolteni: http://www.opengl.org/resources/libraries/glut/
 #include <GL/glut.h>
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Innentol modosithatod...
 
 #define PI 3.141592653589793238462643383
 #define NONE            0
@@ -50,7 +46,7 @@
 const int screenWidth = 600;
 const int screenHeight = 600;
 const uint numBricks = 6;
-const int num_objects = 2+numBricks+1+1+1;//numWalls,numBricks,player,ball,n+1
+const int num_objects = 2 + numBricks + 1 + 1 + 1; //numWalls,numBricks,player,ball,n+1
 const float radius = 1;
 const int detail = 15;
 const float corner_radius = 0.2f;
@@ -58,167 +54,190 @@ const float corner_radius = 0.2f;
 int index = 0;
 
 struct Color {
-   float r, g, b;
-   Color( ) {
-	r = g = b = 0;
-   }
-   Color(float r0, float g0, float b0) {
-	r = r0; g = g0; b = b0;
-   }
+    float r, g, b;
+
+    Color() {
+        r = g = b = 0;
+    }
+
+    Color(float r0, float g0, float b0) {
+        r = r0;
+        g = g0;
+        b = b0;
+    }
 };
 
 struct Margin {
-    Margin(){}
-    Margin(float x, float y, float w, float h) : x0(x), y0(y), w0(w), h0(h){ }
-    float x0,y0,w0,h0;
+    Margin() {}
+
+    Margin(float x, float y, float w, float h) : x0(x), y0(y), w0(w), h0(h) {}
+
+    float x0, y0, w0, h0;
 } margins[4];
 
-void drawCircle(float x, float y, float radius, Color color = Color(1,1,1)) {
-    glColor3f(color.r,color.g,color.b);
+void drawCircle(float x, float y, float radius, Color color = Color(1, 1, 1)) {
+    glColor3f(color.r, color.g, color.b);
     glBegin(GL_TRIANGLE_FAN);
-    for(int i=0;i<=detail;i++){
-        float tmp = PI/180.0f * 360.0f * ( (float) i / (float) detail );
-        glVertex2f(x+radius+radius * (float) cos(tmp),y+radius+radius * (float) sin(tmp));
+    for (int i = 0; i <= detail; i++) {
+        float tmp = PI / 180.0f * 360.0f * ((float) i / (float) detail);
+        glVertex2f(x + radius + radius * (float) cos(tmp), y + radius + radius * (float) sin(tmp));
     }
     glEnd();
 }
 
-void drawRectangle(float x, float y, float w, float h, Color color = Color(1,1,1)) {
-    glColor3f(color.r,color.g,color.b);
+void drawRectangle(float x, float y, float w, float h, Color color = Color(1, 1, 1)) {
+    glColor3f(color.r, color.g, color.b);
     glBegin(GL_TRIANGLES);
-    glVertex2f( x+0.0f,   y+0.0f);
-    glVertex2f( x+w   ,   y+0.0f);
-    glVertex2f( x+w   ,   y+h);
+    glVertex2f(x + 0.0f, y + 0.0f);
+    glVertex2f(x + w, y + 0.0f);
+    glVertex2f(x + w, y + h);
 
-    glVertex2f( x+0.0f,   y+h);
-    glVertex2f( x+0.0f,   y+0.0f);
-    glVertex2f( x+w   ,   y+h);
+    glVertex2f(x + 0.0f, y + h);
+    glVertex2f(x + 0.0f, y + 0.0f);
+    glVertex2f(x + w, y + h);
     glEnd();
 }
 
 bool collidingCircles(float x1, float y1, float radius1, float x2, float y2, float radius2) {
-    float dx = x2 - x1; float dy = y2 - y1;
+    float dx = x2 - x1;
+    float dy = y2 - y1;
     float szumma = radius1 + radius2;
 
-    if ( ( dx * dx )  + ( dy * dy ) < szumma * szumma )
+    if ((dx * dx) + (dy * dy) < szumma * szumma)
         return true;
     else
         return false;
 }
 
 uint collidingAABB(float x, float y, float w, float h, float x0, float y0, float w0, float h0) {
-    if(x0+corner_radius<x+w && x0+w0-corner_radius>x && ((y0+h0>y) && (y0<y+h))){
+    if (x0 + corner_radius < x + w && x0 + w0 - corner_radius > x && ((y0 + h0 > y) && (y0 < y + h))) {
         return HORIZONTALLY;
-    } else if(y0+corner_radius<y && y0+h0-corner_radius>y && ((x0<x+w) && (x0+w0>x))){
+    } else if (y0 + corner_radius < y && y0 + h0 - corner_radius > y && ((x0 < x + w) && (x0 + w0 > x))) {
         return VERTICALLY;
     }
     return NONE;
 }
 
-class Object{
+class Object {
 public:
-    Object(){}
+    Object() {}
+
     Object(float x, float y, float w, float h, Color color, uint type, float vx = 0, float vy = 0)
-        : x(x), y(y), w(w), h(h), color(color), type(type), vx(vx), vy(vy) {
+            : x(x), y(y), w(w), h(h), color(color), type(type), vx(vx), vy(vy) {
     }
 
-    float getX(){return x;} float getY(){return y;}
-    float getW(){return w;} float getH(){return h;}
-    uint getType(){return type;}
+    float getX() { return x; }
+
+    float getY() { return y; }
+
+    float getW() { return w; }
+
+    float getH() { return h; }
+
+    uint getType() { return type; }
 
     void force(float valueX, float valueY) {
-        vx *= sqrt(valueX);     vy *= sqrt(valueY);
+        vx *= sqrt(valueX);
+        vy *= sqrt(valueY);
     }
 
     void move(float dx, float dy) {
-        x += dx;    y += dy;
+        x += dx;
+        y += dy;
     }
 
     void update(float dt) {
-        move(vx*dt,vy*dt);
+        move(vx * dt, vy * dt);
     }
 
-    void draw(){
-        if(type == BALL) {
-            drawCircle(x,y,radius,color);
+    void draw() {
+        if (type == BALL) {
+            drawCircle(x, y, radius, color);
             return;
         }
-        drawRectangle(x,y,w,h,color);
+        drawRectangle(x, y, w, h, color);
     }
 
-    void Intersects(float dt,Object *obj) {
+    void Intersects(float dt, Object *obj) {
         if (type == BRICK && obj->getType() == BRICK)
             return;
 
-        float x0 = obj->getX(); float y0 = obj->getY();
-        float w0 = obj->getW(); float h0 = obj->getH();
+        float x0 = obj->getX();
+        float y0 = obj->getY();
+        float w0 = obj->getW();
+        float h0 = obj->getH();
 
-        uint side = collidingAABB(x,y,w,h,x0,y0,w0,h0);
+        uint side = collidingAABB(x, y, w, h, x0, y0, w0, h0);
 
-        if (type == BALL && obj->getType() == PLAYER ) {
+        if (type == BALL && obj->getType() == PLAYER) {
 
-            if(side == HORIZONTALLY) {
-                force(1.2f,1.2f);
-                move(-(vx+x0+w0)*dt,-vy*dt);
-                vy *=-1;
-            } else if(side == VERTICALLY) {
-                force(1.2f,1.2f);
-                move(-(vx+x0+w0)*dt,-vy*dt);
-                vx *=-1;
-            } else if(side == NONE) {
+            if (side == HORIZONTALLY) {
+                force(1.2f, 1.2f);
+                move(-(vx + x0 + w0) * dt, -vy * dt);
+                vy *= -1;
+            } else if (side == VERTICALLY) {
+                force(1.2f, 1.2f);
+                move(-(vx + x0 + w0) * dt, -vy * dt);
+                vx *= -1;
+            } else if (side == NONE) {
                 float r = radius;
-                float cx = x+r; float cy = y+r;
+                float cx = x + r;
+                float cy = y + r;
 
-                if(collidingCircles(cx,cy,r,x0,y0+h0,corner_radius)) {
-                    force(1.2f,1.2f);
-                    move(-(vx+x0+w0)*dt,-vy*dt);
+                if (collidingCircles(cx, cy, r, x0, y0 + h0, corner_radius)) {
+                    force(1.2f, 1.2f);
+                    move(-(vx + x0 + w0) * dt, -vy * dt);
                     float tmp = vx;
-                    vx = vy;    vy = tmp;
-                } else if(collidingCircles(cx,cy,r,x0+w0,y0+h0,corner_radius)) {
-                    force(1.2f,1.2f);
-                    move(-(vx+x0+w0)*dt,-vy*dt);
+                    vx = vy;
+                    vy = tmp;
+                } else if (collidingCircles(cx, cy, r, x0 + w0, y0 + h0, corner_radius)) {
+                    force(1.2f, 1.2f);
+                    move(-(vx + x0 + w0) * dt, -vy * dt);
                     float tmp = vx;
-                    vx = -vy;   vy = tmp;
-                } else if(collidingCircles(cx,cy,r,x0,y0,corner_radius)) {
-                    force(1.2f,1.2f);
-                    move(-(vx+x0+w0)*dt,-vy*dt);
+                    vx = -vy;
+                    vy = tmp;
+                } else if (collidingCircles(cx, cy, r, x0, y0, corner_radius)) {
+                    force(1.2f, 1.2f);
+                    move(-(vx + x0 + w0) * dt, -vy * dt);
                     float tmp = vx;
-                    vx = -vy;   vy = -tmp;
-                } else if(collidingCircles(cx,cy,r,x0+w0,y0,corner_radius)) {
-                    force(1.2f,1.2f);
-                    move(-(vx+x0+w0)*dt,-vy*dt);
+                    vx = -vy;
+                    vy = -tmp;
+                } else if (collidingCircles(cx, cy, r, x0 + w0, y0, corner_radius)) {
+                    force(1.2f, 1.2f);
+                    move(-(vx + x0 + w0) * dt, -vy * dt);
                     float tmp = vx;
-                    vx = -vy;   vy = tmp;
+                    vx = -vy;
+                    vy = tmp;
                 }
             }
             return;
         }
 
-        if(side !=NONE) {
-            if( type == BALL && obj->getType() == BRICK) {
-                move(-vx*dt,-vy*dt);
+        if (side != NONE) {
+            if (type == BALL && obj->getType() == BRICK) {
+                move(-vx * dt, -vy * dt);
                 float tmpX = vx;
-                float m = radius*radius*PI;
-                float M = w*h;
+                float m = radius * radius * PI;
+                float M = w * h;
 
-                obj->vx       +=  tmpX*(M-m)/(M+m);
-                vx             =  -1.0f * m/M*(tmpX+obj->vx);
+                obj->vx += tmpX * (M - m) / (M + m);
+                vx = -1.0f * m / M * (tmpX + obj->vx);
 
-                if(side == HORIZONTALLY)
-                    vy *=-1;
-            } else if(side == HORIZONTALLY) {
-                vy *=-1;
-            } else if(side == VERTICALLY) {
-                vx *=-1;
+                if (side == HORIZONTALLY)
+                    vy *= -1;
+            } else if (side == HORIZONTALLY) {
+                vy *= -1;
+            } else if (side == VERTICALLY) {
+                vx *= -1;
             }
         }
     }
 
 private:
-    float x,y,w,h;
+    float x, y, w, h;
     Color color;
     uint type;
-    float vx,vy;
+    float vx, vy;
 } objects[num_objects];
 
 class Room {
@@ -227,16 +246,16 @@ public:
     }
 
     void update(float dt) {
-        for(int i=0; i<num_objects-1;i++) {
-            for(int j=i+1;j<num_objects; j++)
-                objects[i].Intersects(dt,&objects[j]);
+        for (int i = 0; i < num_objects - 1; i++) {
+            for (int j = i + 1; j < num_objects; j++)
+                objects[i].Intersects(dt, &objects[j]);
 
             objects[i].update(dt);
         }
     }
 
     void draw() {
-        for(int i=0; i<num_objects;i++)
+        for (int i = 0; i < num_objects; i++)
             objects[i].draw();
     }
 
@@ -246,81 +265,88 @@ public:
 
 } room;
 
-Object* findObject(uint type) {
-    for(int i=0; i<num_objects;i++)
-        if(objects[i].getType() == type)
+Object *findObject(uint type) {
+    for (int i = 0; i < num_objects; i++)
+        if (objects[i].getType() == type)
             return &objects[i];
 
     return NULL;
 }
 
-void onInitialization( ) {
+void onInitialization() {
     glViewport(0, 0, screenWidth, screenHeight);
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    gluOrtho2D(0,50,0,50);
+    gluOrtho2D(0, 50, 0, 50);
 
-    room.addActor(Object(28,22,2,2,Color(1,1,0),BALL,-10,2));
-    room.addActor(Object(33,28,1,4,Color(0,1,0),PLAYER));
-    room.addActor(Object(5,5,40,1,Color(1,1,1),HORIZONTALLY));
-    room.addActor(Object(5,44,40,1,Color(1,1,1),HORIZONTALLY));
+    room.addActor(Object(28, 22, 2, 2, Color(1, 1, 0), BALL, -10, 2));
+    room.addActor(Object(33, 28, 1, 4, Color(0, 1, 0), PLAYER));
+    room.addActor(Object(5, 5, 40, 1, Color(1, 1, 1), HORIZONTALLY));
+    room.addActor(Object(5, 44, 40, 1, Color(1, 1, 1), HORIZONTALLY));
 
-    margins[0] = Margin(5,5,1,40);    margins[1] = Margin(45,5,1,40);
-    margins[2] = Margin(5,44,40,1);    margins[3] = Margin(5,5,40,1);
+    margins[0] = Margin(5, 5, 1, 40);
+    margins[1] = Margin(45, 5, 1, 40);
+    margins[2] = Margin(5, 44, 40, 1);
+    margins[3] = Margin(5, 5, 40, 1);
 
     float step = 38.0f / numBricks;
-    for(int i=0; i<numBricks;i++) {
-        room.addActor(Object(5,6+i*step,1,step,Color(1,0,0),BRICK));
+    for (size_t i = 0; i < numBricks; i++) {
+        room.addActor(Object(5, 6 + i * step, 1, step, Color(1, 0, 0), BRICK));
     }
 
     glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
 }
 
-void onDisplay( ) {
+void onDisplay() {
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
     room.draw();
     glutSwapBuffers();
 }
 
 void onKeyboard(unsigned char key, int x, int y) {
-    Object* player = findObject(PLAYER);
-    float vx = 0;   float vy = 0;
+    Object *player = findObject(PLAYER);
+    float vx = 0;
+    float vy = 0;
 
-    float xx = player->getX();    float yy = player->getY();
-    float ww = player->getW();    float hh = player->getH();
+    float xx = player->getX();
+    float yy = player->getY();
+    float ww = player->getW();
+    float hh = player->getH();
     bool collided = false;
 
-    if(key=='i')
+    if (key == 'i')
         vy = 1;
-    else if(key=='m')
+    else if (key == 'm')
         vy = -1;
-    else if(key=='j')
+    else if (key == 'j')
         vx = -1;
-    else if(key=='k')
+    else if (key == 'k')
         vx = 1;
 
-    for(int i=0;i<4;i++) {
-        float x0 = margins[i].x0;        float y0 = margins[i].y0;
-        float w0 = margins[i].w0;        float h0 = margins[i].h0;
+    for (int i = 0; i < 4; i++) {
+        float x0 = margins[i].x0;
+        float y0 = margins[i].y0;
+        float w0 = margins[i].w0;
+        float h0 = margins[i].h0;
 
-        if(collidingAABB(xx+vx,yy+vy,ww,hh,x0,y0,w0,h0))
+        if (collidingAABB(xx + vx, yy + vy, ww, hh, x0, y0, w0, h0))
             collided = true;
     }
 
-    if(!collided)   player->move(vx,vy);
+    if (!collided) player->move(vx, vy);
 }
 
 void onMouse(int button, int state, int x, int y) {
 }
 
-void onIdle( ) {
+void onIdle() {
     static float time = 0;
     const float Dt = 0.1;
 
     float newTime = glutGet(GLUT_ELAPSED_TIME) / 1000.0f;
 
-    for(; time < newTime; time+= Dt) {
-        if(time + Dt > newTime){
+    for (; time < newTime; time += Dt) {
+        if (time + Dt > newTime) {
             room.update(newTime - time);
             time = newTime;
             break;
